pyproject.toml
[tool.poetry]
name = "demando"
version = "0.1.0"
description = "App for q&a and votes."
authors = ["Oleh Pykulytsky <pykulytsky@gmail.com>"]
license = "MIT"

readme = 'README.md'

repository = "https://github.com/pykulytsky/demando"
homepage = "https://github.com/pykulytsky/demando"

packages = [
    {include="scripts", from="src"}
]

[tool.poetry.dependencies]
python = "^3.8"
gino = {extras = ["pg", "starlette"], version = "^1.0.1"}
alembic = "1.5.7"
psycopg2 = "2.8.6"
asgi-lifespan = "1.0.1"
attrs = "20.3.0"
autopep8 = "1.5.5"
backcall = "0.2.0"
certifi = "2020.12.5"
chardet = "4.0.0"
click = "8.0.0"
colorama = "0.4.4"
coverage = "5.5"
decorator = "4.4.2"
dnspython = "2.1.0"
email-validator = "1.1.2"
fastapi = "0.63.0"
h11 = "0.12.0"
httpcore = "0.12.3"
httpx = "0.17.0"
idna = "2.10"
iniconfig = "1.1.1"
ipython = "7.22.0"
ipython_genutils = "0.2.0"
jedi = "0.18.0"
Mako = "1.1.4"
MarkupSafe = "1.1.1"
mccabe = "0.6.1"
mixer = "7.1.2"
packaging = "20.9"
parso = "0.8.1"
passlib = "1.7.4"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pluggy = "0.13.1"
prompt-toolkit = "3.0.17"
ptyprocess = "0.7.0"
py = "1.10.0"
pycodestyle = "2.7.0"
pydantic = "1.8.1"
pyflakes = "2.3.0"
Pygments = "2.8.1"
PyJWT = "1.7.1"
pyparsing = "2.4.7"
python-dateutil = "2.8.1"
python-decouple = "3.4"
python-editor = "1.0.4"
requests-mock = "1.8.0"
rfc3986 = "1.4.0"
sentry-sdk = "1.0.0"
shellingham = "1.4.0"
six = "1.15.0"
sniffio = "1.2.0"
SQLAlchemy = "1.3.23"
starlette = "0.13.6"
text-unidecode = "1.3"
toml = "0.10.2"
tqdm = "4.59.0"
traitlets = "^5.0.5"
typer = "0.4.1"
typing-extensions = "3.10.0.0"
urllib3 = "1.26.3"
wcwidth = "0.2.5"
orm = "^0.1.5"
mailjet-rest = "^1.3.4"
pytest-env = "^0.6.2"
websockets = "^10.1"
uvicorn = {extras = ["standard"], version = "^0.17.1"}
hypercorn = "^0.13.2"
black = "^22.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
flake8 = "^3.9.0"
pytest-asyncio = "^0.14.0"
pytest-bug = "^1.0.0"
pytest-cov = "^2.11.1"
pytest-mock = "^3.5.1"
pytest-picked = "^0.4.6"
ipython = "^7.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'scripts.dev:main'
prod = 'scripts.prod:main'
truncate = 'scripts.truncate_db:main'
truncate-test = 'scripts.truncate_test_db:main'
